maintenance

+++ index.js
@@ -57,7 +57,9 @@ client.on('message', message => { // Activities code below when a message is sen
         if (foundInText == false) return; //Checks if the message contains mentions and if it does
     }
 
-    if (foundInText) message.delete();
+    if (maintain == 'on') {
+        if (foundInText) message.delete();
+    }
 
 });
